requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
Click==7.0
lxml==4.5.0
numpy==1.18.1
pandas==1.0.1
pathspec==0.7.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2019.3
regex==2020.2.20
six==1.14.0
soupsieve==2.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
